Moves image tests
