+++ packages/image/__tests__/android/placeholder.android.test.js
