+++ packages/image/__tests__/image.native.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import renderer from "react-test-renderer";
-import Image from "../image";
+import Image from "../src/image";
 
 export default () => {
   it("renders correctly", () => {
