+++ packages/image/__tests__/ios/jest.config.js
