+++ packages/image/__tests__/web/placeholder.web.test.js
